feat:订单列表过渡

+++ app/src/main/java/com/jcs/where/features/gourmet/takeaway/order/TakeawayOrderDetailActivity.kt
@@ -59,7 +59,7 @@ class TakeawayOrderDetailActivity : BaseMvpActivity<TakeawayOrderDetailPresenter
 
     override fun initData() {
         presenter = TakeawayOrderDetailPresenter(this)
-        presenter.getDetail(orderId)
+        presenter.getDetail(orderId.toInt())
     }
 
     override fun bindListener() {
